Move HTTP proxy tests in a separate class
