+++ src/test/java/io/vertx/core/http/Http1xTest.java
@@ -46,7 +46,6 @@ import java.util.stream.IntStream;
 import static io.vertx.test.core.TestUtils.*;
 
 /**
- *
  * @author <a href="http://tfox.org">Tim Fox</a>
  * @author <a href="mailto:nscavell@redhat.com">Nick Scavelli</a>
  */
@@ -2967,136 +2966,6 @@ public class Http1xTest extends HttpTest {
     });
     req.putHeader("the_header", TestUtils.randomAlphaString(10000));
     req.sendHead();
-    await();
-  }
-
-  @Test
-  public void testHttpProxyRequest() throws Exception {
-    startProxy(null, ProxyType.HTTP);
-    client.close();
-    client = vertx.createHttpClient(new HttpClientOptions()
-        .setProxyOptions(new ProxyOptions().setType(ProxyType.HTTP).setHost("localhost").setPort(proxy.getPort())));
-    testHttpProxyRequest2(client.get(DEFAULT_HTTP_PORT, DEFAULT_HTTP_HOST, "/"));
-  }
-
-  @Test
-  public void testHttpProxyRequestOverrideClientSsl() throws Exception {
-    startProxy(null, ProxyType.HTTP);
-    client.close();
-    client = vertx.createHttpClient(new HttpClientOptions()
-        .setSsl(true).setProxyOptions(new ProxyOptions().setType(ProxyType.HTTP).setHost("localhost").setPort(proxy.getPort())));
-    testHttpProxyRequest2(client.get(new RequestOptions().setSsl(false).setHost("localhost").setPort(8080)));
-  }
-
-  private void testHttpProxyRequest2(HttpClientRequest clientReq) throws Exception {
-    server.requestHandler(req -> {
-      req.response().end();
-    });
-
-    server.listen(onSuccess(s -> {
-      clientReq.handler(resp -> {
-        assertEquals(200, resp.statusCode());
-        assertNotNull("request did not go through proxy", proxy.getLastUri());
-        assertEquals("Host header doesn't contain target host", "localhost:8080", proxy.getLastRequestHeaders().get("Host"));
-        testComplete();
-      });
-      clientReq.exceptionHandler(this::fail);
-      clientReq.end();
-    }));
-    await();
-  }
-
-  @Test
-  public void testHttpProxyRequestAuth() throws Exception {
-    startProxy("user", ProxyType.HTTP);
-
-    client.close();
-    client = vertx.createHttpClient(new HttpClientOptions()
-        .setProxyOptions(new ProxyOptions().setType(ProxyType.HTTP).setHost("localhost").setPort(proxy.getPort())
-            .setUsername("user").setPassword("user")));
-
-    server.requestHandler(req -> {
-      req.response().end();
-    });
-
-    server.listen(onSuccess(s -> {
-      client.get(DEFAULT_HTTP_PORT, DEFAULT_HTTP_HOST, "/", resp -> {
-        assertEquals(200, resp.statusCode());
-        assertNotNull("request did not go through proxy", proxy.getLastUri());
-        assertEquals("Host header doesn't contain target host", "localhost:8080", proxy.getLastRequestHeaders().get("Host"));
-        testComplete();
-      }).exceptionHandler(th -> fail(th)).end();
-    }));
-    await();
-  }
-
-  @Test
-  public void testHttpProxyFtpRequest() throws Exception {
-    startProxy(null, ProxyType.HTTP);
-    client.close();
-    client = vertx.createHttpClient(new HttpClientOptions()
-        .setProxyOptions(new ProxyOptions().setType(ProxyType.HTTP).setHost("localhost").setPort(proxy.getPort())));
-    final String url = "ftp://ftp.gnu.org/gnu/";
-    proxy.setForceUri("http://localhost:8080/");
-    HttpClientRequest clientReq = client.getAbs(url);
-    server.requestHandler(req -> {
-      req.response().end();
-    });
-
-    server.listen(onSuccess(s -> {
-      clientReq.handler(resp -> {
-        assertEquals(200, resp.statusCode());
-        assertEquals("request did sent the expected url", url, proxy.getLastUri());
-        testComplete();
-      });
-      clientReq.exceptionHandler(this::fail);
-      clientReq.end();
-    }));
-    await();
-  }
-
-  @Test
-  public void testHttpSocksProxyRequest() throws Exception {
-    startProxy(null, ProxyType.SOCKS5);
-
-    client.close();
-    client = vertx.createHttpClient(new HttpClientOptions()
-        .setProxyOptions(new ProxyOptions().setType(ProxyType.SOCKS5).setHost("localhost").setPort(proxy.getPort())));
-
-    server.requestHandler(req -> {
-      req.response().end();
-    });
-
-    server.listen(onSuccess(s -> {
-      client.get(DEFAULT_HTTP_PORT, DEFAULT_HTTP_HOST, "/", resp -> {
-        assertEquals(200, resp.statusCode());
-        assertNotNull("request did not go through proxy", proxy.getLastUri());
-        testComplete();
-      }).exceptionHandler(th -> fail(th)).end();
-    }));
-    await();
-  }
-
-  @Test
-  public void testHttpSocksProxyRequestAuth() throws Exception {
-    startProxy("user", ProxyType.SOCKS5);
-
-    client.close();
-    client = vertx.createHttpClient(new HttpClientOptions()
-        .setProxyOptions(new ProxyOptions().setType(ProxyType.SOCKS5).setHost("localhost").setPort(proxy.getPort())
-            .setUsername("user").setPassword("user")));
-
-    server.requestHandler(req -> {
-      req.response().end();
-    });
-
-    server.listen(onSuccess(s -> {
-      client.get(DEFAULT_HTTP_PORT, DEFAULT_HTTP_HOST, "/", resp -> {
-        assertEquals(200, resp.statusCode());
-        assertNotNull("request did not go through proxy", proxy.getLastUri());
-        testComplete();
-      }).exceptionHandler(th -> fail(th)).end();
-    }));
     await();
   }
